+++ CHANGELOG.md
@@ -8,3 +8,8 @@
 # v0.1.1 | Throw passport anywhere patch
 
 - Added PassportPatch.cs and registerd it as GlobalPatch in Plugin.cs
+
+# v0.1.2 | Bags for Everyone
+
+- Added BagsForEveryoneModule and BagsForEveryonePatch
+- BagsForEveryonePatch adds a postfix to SingleItemSpawner.TrySpawnItems that checks if the spawner is the first one in the biome, and if so, it spawns extra bags based on the player count.